project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "openai-whisper",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.2",
    "sounddevice>=0.5.2",
]
